Update main.c
In this contribution, I added a condition to the while loop that checks if the user entered the letter 'q'. If the user enters 'q', the program will break out of the loop and print a message indicating that the program is exiting. This change makes the program more interactive and responsive to user input.